requirements.txt
flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0
requests==2.31.0
google-cloud-firestore==2.16.0
google-generativeai==0.8.4
python-dotenv==1.0.1